Renamed and reoganized project for new repository
