+++ .gitignore
@@ -1,3 +1,3 @@
-direct_todo/directtodo
+todo_django/directtodo
 *.pyc
 *.swp
