+++ todo_django/templates/index.html
@@ -4,7 +4,7 @@
     <div class="wrapper">    
         <div class="row">
             <div class="span6">            
-                <h1>Direct Todo</h1>
+                <h1>Todo Django</h1>
                 <p>A Django powered to do app for the IndyPy Web Shootout</p>
             </div>          
             <div class="span6">
