+++ todo_django/templates/site_base.html
@@ -64,7 +64,7 @@
             <div class="row">
                 <div class="span3" id="logo-block">
                     <a href="/">
-                       <h3 class="pull-left">Django Todo</h3>
+                       <h3 class="pull-left">Direct Employers</h3>
                     </a>                    
                 </div>  
                 <div class="span8">
